% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_corpus.R
\name{build_corpus}
\alias{build_corpus}
\title{build_corpus}
\usage{
build_corpus(x)
}
\arguments{
\item{x}{A list of plain text files}
}
\value{
A clean {\link{VCorpus}} object, ready to be modeled
}
\description{
Builds and cleans the corpus to be modeled
}
