% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhtk_build_corpus.R
\name{dhtk_build_corpus}
\alias{dhtk_build_corpus}
\title{dhtk_build_corpus}
\usage{
dhtk_build_corpus(x)
}
\arguments{
\item{x}{A list of plain text files}
}
\value{
A clean {\link{VCorpus}} object, ready to be modeled
}
\description{
Builds and cleans the corpus to be modeled
}
