% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhtk_build_json.R
\name{dhtk_build_json}
\alias{dhtk_build_json}
\title{dhtk_build_json}
\usage{
dhtk_build_json(x, y, z)
}
\arguments{
\item{x}{a topic model object from topicmodels package}

\item{y}{a corpus object from tm package}

\item{z}{a document term matrix object from tm package}
}
\value{
json object
}
\description{
Builds a json object for online visualizations
}
