% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_web.R
\name{get_web}
\alias{get_web}
\title{get_web}
\usage{
get_web(x, y)
}
\arguments{
\item{x}{a url}

\item{y}{a css selector}
}
\value{
character string
}
\description{
one-liner for scraping text data from the web
}
\examples{
get_web("http://classics.mit.edu/Homer/iliad.23.xxiii.html", "blockquote")
}
